+++ src/EntityGraphQL/Schema/FieldExtensions/Filter/FilterArgs.cs
@@ -2,6 +2,6 @@ namespace EntityGraphQL.Schema.FieldExtensions
 {
     public class FilterArgs<T>
     {
-        public EntityQueryType<T> Filter { get; set; } = ArgumentHelper.EntityQuery<T>();
+        public EntityQueryType<T>? Filter { get; set; } = ArgumentHelper.EntityQuery<T>();
     }
 }