+++ src/EntityGraphQL/Schema/FieldExtensions/Sorting/SortArgs.cs
@@ -5,7 +5,7 @@ namespace EntityGraphQL.Schema.FieldExtensions
 {
     public class SortInput<T> where T : notnull
     {
-        public List<T> Sort { get; set; } = new List<T>();
+        public List<T>? Sort { get; set; }
     }
 
     public enum SortDirectionEnum