+++ src/tests/EntityGraphQL.Tests/SortTests/SortTests.cs
@@ -169,7 +169,7 @@ namespace EntityGraphQL.Tests
             var schemaType = schema.Type("PeopleSortInput");
             var fields = schemaType.GetFields().ToList();
             Assert.Equal(11, fields.Count);
-            Assert.Contains("people(sort: [PeopleSortInput!]! = []): [Person!]", schema.ToGraphQLSchemaString());
+            Assert.Contains("people(sort: [PeopleSortInput!] = [{ height: ASC }]): [Person!]", schema.ToGraphQLSchemaString());
         }
         [Fact]
         public void SupportUseSortOnNonRoot()